Implement TCP upstream facilities
The TCP layer is the implementation layer responsible for
facilitating the deployment of TCP servers via the abstractions
provided in the policy layer. It is entirely contained in the
Networking::TCP namespace, and it includes classes to receive and
handle TCP connections (calling user code, of course). The client
code (provided via class TCPClient) will be added in a future
update.