Reorder arguments, remove destructor
The biggest thing this patch does is remove the destructor for
TCPListener. The socket is now wrapped in a shared_ptr with a
custom deleter which closes the socket. This allows us to use the
default move/copy constructor/assignment operator without danger of
accidentally closing the socket or double-closing it.

Also reorder the arguments in the constructor of TCPListener to be
consistent with TCPClient.