Don't throw if client fails TLS handshake
Previously, if the client failed the TLS handshake, the TLSListener
would throw an exception, likely causing the server to fail. This
has obvious security implications. We now allow the user to set the
behavior in this scenario in the TLSListener::Builder (and via the
constructor). The choices are:

  1. The listener puts a message into the log containing the OpenSSL
     error trace and the IP address of the client.
  2. The listener throws a custom exception,
     'ClientFailedTLSHandshakeException.' Obviously named, this
     allows the user to interrupt the normal flow of logic so that
     they can, for example, add that IP to a ban list, etc.

In the future, it is likely that we will expand the list of options
to include an 'allow insecure connections' option, which will revert
to normal TCP operation in this scenario.