Require OpenSSL v1.1.0+
In version 1.0.2 or earlier, the user must explicitly initialize the
OpenSSL library. This explicit initialization is very difficult to
manage in an object-oriented context, even using Modern C++. In
OpenSSL version 1.1.0 or greater, the library initializes itself
automatically. To save complexity and prevent the possibility of
unknown state, we now require OpenSSL version 1.1.0 or greater to
be installed.