Separate NetworkAddress from NetworkHost
Previously, all logic for manipulating IP addresses was contained
in the NetworkHost class. We abstract this away to a separate class,
NetworkAddress, in preparation for adding hostname resolution
functionality to the NetworkHost class. It is believed that these
are different responsibilities. Some changes needed to be made to
the template code to allow them to work with the new class.