Provide template specializations for NetworkHost
Previously, the library would not compile if the library templates
were initialized with NetworkHost as a template parameter. This was
due to insufficient specialization. This patch provides that
specialization, as well as empty implementations of the missing
NetworkHost functions, so that the library compiles. NetworkHost has
not yet been fully implemented.