Allow masking of SIGPIPE, fix bug
Previously, the TCPListener did not mask SIGPIPE, which would cause
the server to terminate if, for example, the client closed the
connection before the server finished the TLS handshake. This is
obviously a security issue. Add 'maskSigPipe' paramter to
TCPListener ctor. If true, the TCPListener permanently masks
SIGPIPE. The caller can set this to false if it wishes to provide
its own override.

Also fix a bug that caused TCPListener ctor to bind to the wrong
address.